backend/pyproject.toml
[project]
name = "envctl"
version = "0.1.0"
description = "Internal developer platform API for managing environments and deployments."
authors = [
    { name = "Andreas Schubert" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.2,<0.122.0",
    "uvicorn[standard]>=0.38.0,<0.39.0",
    "sqlalchemy>=2.0.44,<3.0.0",
    "alembic>=1.17.2,<2.0.0",
    "psycopg2-binary>=2.9.9,<3.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "argon2-cffi (>=25.1.0,<26.0.0)"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "httpx>=0.27.0,<0.28.0",
    "ruff>=0.7.0,<0.8.0",
    "black>=24.0.0,<25.0.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
[dependency-groups]
dev = [
    "pytest (>=8.0.0,<9.0.0)",
    "httpx (>=0.27.0,<0.28.0)",
    "ruff (>=0.7.0,<0.8.0)",
    "black (>=24.0.0,<25.0.0)"
]

[tool.pytest.ini_options]
pythonpath = "."
